refactor(markdown): split parser into submodules and centralize quarantine class
The 1139-line markdown.rs bundled the State event-stream driver, the
Frame stack data type, and the post-parse Pandoc/attribute rewrite.
Split along the natural seams:

- infrastructure/parser/markdown/frame.rs (118 lines): Frame enum + the
  find_enclosing_table helper that walks the stack past TableHead /
  TableRow wrappers.
- infrastructure/parser/markdown/post_parse.rs (65 lines): the
  rewrite_pandoc_and_attribute_paragraphs pass that converts
  Pandoc/attribute paragraphs into UnknownExtension blocks via the
  shared extension_classifier.
- infrastructure/parser/markdown/mod.rs (1006 lines): the State driver
  and the public parse_markdown_page entry point.

Behavior unchanged — pure relocation; the workspace test suite stays
green.

Side-effect cleanup: introduce QUARANTINED_HTML_CLASS in
infrastructure/render/html.rs as the single source for the quarantine
CSS class string (was hardcoded as 'quarantined-html' at two emit
sites). CONTEXT.md is corrected to match the code — the previous
glossary entry named 'adoc-quarantined-html' but no consumer ever saw
that string. Renaming the class is now a single-constant edit.
